Split fuentes primarias out of the IPC monolith
  electron/ipc.ts     2358-2856 -> electron/ipc/primarySources.ts     (59 handlers)
  electron/preload.ts   498-601 -> electron/preload/primarySources.ts (59 bindings)
  shared/types.ts     7856-8046 -> shared/api/primarySources.ts       (59 signatures)

This domain declares its types with inline imports —
`Promise<import('./primarySourcesTypes').PrimarySourceDossier>` — and those paths
were relative to shared/types.ts, so from shared/api/ they silently resolved to
nothing. The visible symptom was not an error in the contract but nine implicit-any
errors in the renderer views, because every affected method's return type collapsed.
The extraction now rewrites those paths, and typecheck is what caught it.

Nine primary-sources tests read the three hot files by path to match source text.
Rather than repoint them at new filenames, readSource() in the channel census takes
three sentinels — '@main', '@bridge', '@api' — that stand for "wherever that surface
lives now". Swapping a path for a sentinel keeps each assertion and its position in a
list intact, so the migration is mechanical, and it is the pattern for the ~48
remaining tests with the same habit.

electron/ipc.ts is 3,906 lines (from 6,013 — down 35%), electron/preload.ts 1,326
(from 1,862 — down 29%), shared/types.ts 8,788 (from 9,553). 371 channels live in a
domain module.

Verified: 1,425 tests, lint and typecheck clean, contract test green.
