+++ scripts/test-primary-sources-files-viewer.mjs
@@ -6,23 +6,24 @@ import { createRequire } from 'node:module';
 import os from 'node:os';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { readSource } from './ipc-channel-census.mjs';
 
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
 const require = createRequire(import.meta.url);
 
 if (!process.argv.includes('--electron-primary-sources-files-test')) {
-  const indexHtml = fs.readFileSync(path.join(repoRoot, 'index.html'), 'utf8');
+  const indexHtml = readSource('index.html');
   const [view, archiveView, protocol, fileRepo, shared, preload, ipc, schema, sync] = [
     'src/views/PrimarySourceDossierView.tsx',
     'src/views/PrimarySourcesArchiveView.tsx',
     'electron/archiveProtocol.ts',
     'electron/db/archiveFilesRepo.ts',
-    'shared/types.ts',
-    'electron/preload.ts',
-    'electron/ipc.ts',
+    '@api',
+    '@bridge',
+    '@main',
     'electron/db/migrations.ts',
     'electron/db/syncTables.ts',
-  ].map((file) => fs.readFileSync(path.join(repoRoot, file), 'utf8'));
+  ].map((file) => readSource(file));
   for (const marker of [
     'primary-source-image-viewer',
     'primary-source-pdf-viewer',
